registration bug fixes
