+++ mars_behavior/src/mars_behavior/perception.py
@@ -1,6 +1,7 @@
 import rospy
 
 from mars_msgs.srv import RegistrationSrv
+from geometry_msgs.msg import Pose
 
 
 class PerceptionInterface:
@@ -10,8 +11,8 @@ class PerceptionInterface:
             f"/{self.ns_}/registration", RegistrationSrv
         )
 
-    def run_object_registration(self, mesh_name: str):
+    def run_object_registration(self, mesh_name: str, initial_tf: Pose = Pose()):
         rospy.set_param(f"/{self.ns_}/detect_class_names", mesh_name)
         rospy.sleep(0.1)
-        self.registration_srv(mesh_name)
+        self.registration_srv(mesh_name=mesh_name,init_tf=initial_tf)    
         rospy.sleep(3.0)  # time to converge
