+++ mars_msgs/srv/RegistrationSrv.srv
@@ -1,3 +1,3 @@
 string mesh_name 
-geometry_msgs/Transform init_tf
+geometry_msgs/Pose init_tf
 ---
