+++ mars_perception/CMakeLists.txt
@@ -20,8 +20,8 @@ find_package(catkin REQUIRED COMPONENTS
 
 find_package(PCL REQUIRED) # This includes all modules
 find_package(Eigen3 REQUIRED)
-find_package(Open3D REQUIRED HINTS ${CMAKE_INSTALL_PREFIX}/lib/CMake)
-find_package(Ceres REQUIRED)
+# find_package(Open3D REQUIRED HINTS ${CMAKE_INSTALL_PREFIX}/lib/CMake)
+# find_package(Ceres REQUIRED)
 
 
 find_package(realsense2 2.50.0)
@@ -59,12 +59,15 @@ include_directories(
   ${catkin_INCLUDE_DIRS}
   ${EIGEN3_INCLUDE_DIRS}
   ${PCL_INCLUDE_DIRS}
-  ${Open3D_INCLUDE_DIRS} 
-  ${CERES_INCLUDE_DIRS}
+  # ${Open3D_INCLUDE_DIRS} 
+  # ${CERES_INCLUDE_DIRS}
 )
 
 
-link_directories(${PCL_LIBRARY_DIRS} ${Open3D_LIBRARY_DIRS}) 
+link_directories(
+  ${PCL_LIBRARY_DIRS}
+  # ${Open3D_LIBRARY_DIRS}
+) 
 add_definitions(${PCL_DEFINITIONS})
 
 add_executable(${PROJECT_NAME}_processor src/pc_processor_node.cpp)
@@ -74,15 +77,20 @@ target_link_libraries(${PROJECT_NAME}_processor
   ${catkin_LIBRARIES}
   ${PCL_LIBRARIES}
 )
-add_executable(${PROJECT_NAME}_registration src/registration_server.cpp src/registration/icp.cpp src/mesh_sampling.cpp src/common.cpp src/registration/constrained_icp.cpp)
+add_executable(${PROJECT_NAME}_registration 
+  src/registration_server.cpp src/registration/icp.cpp 
+  src/mesh_sampling.cpp 
+  src/common.cpp 
+  # src/registration/constrained_icp.cpp
+)
 set_target_properties(${PROJECT_NAME}_registration PROPERTIES OUTPUT_NAME registration_server PREFIX "")
 add_dependencies(${PROJECT_NAME}_registration ${${PROJECT_NAME}_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
 target_link_libraries(${PROJECT_NAME}_registration
   ${catkin_LIBRARIES}
   ${PCL_LIBRARIES}
   ${Eigen3_LIBRARIES}
-  ${Open3D_LIBRARIES}
-  ${CERES_LIBRARIES}
+  # ${Open3D_LIBRARIES}
+  # ${CERES_LIBRARIES}
 )
 
 add_executable(${PROJECT_NAME}_registration_test src/registration_test.cpp)
