+++ mars_perception/src/common.cpp
@@ -43,13 +43,3 @@ bool MeshUtil::update_mesh(std::string name, Eigen::Affine3d tf)
     }
     return true;
 }
-
-void pcl_to_open3d(PointCloudPtr pc_ptr, open3d::geometry::PointCloud& open3d_p) {
-    size_t pc_size = pc_ptr->size();
-    open3d_p.points_.resize(pc_size);
-    for(int i = 0; i < pc_ptr->size(); i++) {
-        Eigen::Vector3d v;
-        v << pc_ptr->points[i].x, pc_ptr->points[i].y, pc_ptr->points[i].z;
-        open3d_p.points_[i] = v;
-    }
-}
